[feat] #52 Add bind() logic
